created structure for the project
